labs-setup/oci-logs-collection-scripts/requirements.txt
oci>=2.128.0
python-dotenv>=1.0.0
requests>=2.28.0
prometheus_client>=0.20.0
